Attempt to fix out-of-sequence Game updates
Add more precision back into Game#version so that we can see if it helps
the occasional issue I've seen where my move shows up and then is
quickly erased again. I believe this is due to the complex nature of
having both the render result and the broadcast result for the user
making the move. Thus, if the broadcast is out of sequence and is within
the same second then it would replace the render result, which could
slide the board state back a bit.

+++ app/models/game.rb
@@ -125,7 +125,7 @@ class Game < ApplicationRecord # rubocop:disable Metrics/ClassLength
   def display_id = "##{id.to_s.rjust(self.class.display_id_width, "0")}"
 
   def version
-    @version ||= Time.now.to_i
+    @version ||= Time.now.to_f
   end
 
   # :reek:TooManyStatements
